requirements.txt
langgraph
langchain
langchain_community
langchain_openai
langchain_google_genai
google-generativeai
litellm
pydantic
pdfplumber
easyocr
pdf2image
pillow
numpy
pandas
fastapi
uvicorn
requests
streamlit
watchdog
python-dotenv
pyyaml
langfuse
faiss-cpu